Fixed infinite loop for very large inputs
If the most significant bit of 'x' was 1, the program would be unable to find a value of 'n' to rightshift 'x'. For example, if 'x' is a 32 bit integer, then 'x >> 32' will produce 'x' instead of 0.

Support for long integers has been temporarily removed as logical NOT (!) will always produce a regular integer type.